OrderedCommandParser
Created an OrderedCommandParser and Tests. This parses commands and
matches parameters to the routes parameters in the order they are
specified in the route. So

Route: ls path
Command: ls c:\users\

the path parameter gets set to c:\users\

Moved both of these into there own little folder and had to change the
IParseCommands interface so you can give a parser the routes we
currently know about so the ordered parser can match the parameters in
the command to the route