starting v1.1
Abstracted out the Console.Write and Console.Read into a very basic
ICmdRConsole interface with a basic Console implemetation. this is
blocking atm and will need to be looked at if we are going to implement
a wpf version.

Wrapped the state for the prompt and exit code into a ICmdRState so we
can pass this around so our Actions can modify it on the fly.

Changed the Actions so they take in the ICmdRState and ICmdRConsole so
we can use these within the actions and modify use them as we needed