Adding the ability to parse switches from commands, switches are denoted by /switch
Added a new Register Route which has an action which only requires the param to be passed in and not the CmdR or Console and State params. Which can be used by Modules as these accept the CmdR instance in there constructor and can store that reference so they don't need it passed in for evert method. this make it a little neater.